refactor: getColumnName 함수를 ColumnDefinitionMapper로 이동
